\name{CPL-package}
\alias{CPL-package}
\alias{CPL}
\docType{package}
\title{
  R bindings for Prov-CPL
}
\description{
  Provides basic Prov-CPL functionality around the upload
  and download of JSON documents. Full library support not yet 
  complete. 
}
\details{
  TODO
}
\author{
Jackson Okuhn.

Maintainer: Jackson Okuhn <jpokuhn@gmail.com>
}
\references{
  TODO
}
\keyword{ package }
\seealso{
  TODO
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
